+++ src/components/media/media-list.spec.js
@@ -19,14 +19,14 @@ describe("media list tests", () => {
 
   it("It selects images (media-list)", () => {
     cy.get("#delete_media_button").should("be.disabled");
-    cy.get(".media-list").find("button").eq(0).click();
+    cy.get(".media-list").find("input").eq(0).click();
     cy.get(".media-list").find(".card").should("have.class", "selected");
     cy.get("#delete_media_button").should("not.be.disabled");
   });
 
   it("It opens delete modal (media-list)", () => {
     cy.get("#delete-modal").should("not.exist");
-    cy.get(".media-list").find("button").eq(0).click();
+    cy.get(".media-list").find("input").eq(0).click();
     cy.get("#delete_media_button").click();
     cy.get("#delete-modal").should("exist");
   });